requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
PyPDF2>=3.0.1
pdfplumber>=0.11.0
python-docx>=1.1.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
plotly>=5.20.0
Pillow>=10.2.0
tqdm>=4.66.0
